Implement ejabberd_suspend sockmod type and use it
This is a wrapper around any other module that implements the
sockmod interface. Currently it doesn't do any of the clever
queueing stuff yet, so it is just another layer of indirection.

It cribs heavily off ejabberd_socket.erl

Note that in ejabberd_c2s, the reply stanza to <sleep/> is routed
using the normal mechanism. This is a temporary measure to make
my tests pass. Once I get queueing working, it will have to be sent
in as the last argument to ejabberd_suspend:sleep/3, or it won't
get routed until <wake/> happens (which seems pointless)